requirements.txt
flask>=3.0,<4.0
pyserial>=3.5,<4.0
pytest>=8.0,<9.0
